# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a seed sync server.",
        "title": "Seed Sync API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/cookie-cloud/add-or-update": {
            "post": {
                "description": "添加或更新cookie cloud配置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cookie cloud"
                ],
                "summary": "添加或更新cookie cloud配置",
                "parameters": [
                    {
                        "description": "cookie cloud配置",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.CookieCloudConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/cookie-cloud/get": {
            "get": {
                "description": "获取cookie cloud配置",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cookie cloud"
                ],
                "summary": "获取cookie cloud配置",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "db.CookieCloudConfig": {
            "type": "object",
            "properties": {
                "p2p_password": {
                    "description": "端对端加密密码",
                    "type": "string"
                },
                "sync_cron": {
                    "description": "同步cron表达式",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_key": {
                    "description": "用户KEY",
                    "type": "string"
                }
            }
        }
    }
}